requirements.txt
flake8==3.5.0
ptpython==0.41
PyJWT==1.5.3
pytest-freezegun==0.2.0
pytest==3.4.0
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
snapshottest==0.5.0
